move edit eth navigator
