+++ apps/ledger-live-mobile/src/components/RootNavigator/BaseNavigator.tsx
@@ -72,6 +72,7 @@ import WalletConnectNavigator from "./WalletConnectNavigator";
 import WalletConnectLiveAppNavigator from "./WalletConnectLiveAppNavigator";
 import CustomImageNavigator from "./CustomImageNavigator";
 import ClaimNftNavigator from "./ClaimNftNavigator";
+import EditTransactionNavigator from "../../families/ethereum/EditTransaction/EditTransactionNavigator";
 import PostOnboardingNavigator from "./PostOnboardingNavigator";
 import { readOnlyModeEnabledSelector } from "../../reducers/settings";
 import { hasNoAccountsSelector } from "../../reducers/accounts";
@@ -89,7 +90,6 @@ import {
 } from "../NavigationHeaderCloseButton";
 import { RedirectToRecoverStaxFlowScreen } from "../../screens/Protect/RedirectToRecoverStaxFlow";
 import { RootDrawer, RootDrawerProps } from "../RootDrawer";
-import EditTransactionNavigator from "../../families/ethereum/EditTransactionNavigator";
 
 const Stack = createStackNavigator<BaseNavigatorStackParamList>();
