+++ apps/ledger-live-mobile/src/families/ethereum/EditTransaction/EditTransactionNavigator.tsx
@@ -3,17 +3,17 @@ import { createStackNavigator } from "@react-navigation/stack";
 import { useTranslation } from "react-i18next";
 import { useTheme } from "styled-components/native";
 
-import { ScreenName } from "../../const";
-import { getStackNavigatorConfig } from "../../navigation/navigatorConfig";
-import StepHeader from "../../components/StepHeader";
-import { MethodSelection } from "./EditTransaction/MethodSelection";
-import { EditTransactionParamList } from "./EditTransactionParamList";
-import SelectDevice from "../../screens/SelectDevice";
-import SendConnectDevice from "../../screens/ConnectDevice";
-import SendSummary from "../../screens/SendFunds/04-Summary";
-import SendValidationSuccess from "../../screens/SendFunds/07-ValidationSuccess";
-import SendValidationError from "../../screens/SendFunds/07-ValidationError";
-import { TransactionAlreadyValidatedError } from "./EditTransaction/TransactionAlreadyValidatedError";
+import { ScreenName } from "../../../const";
+import { getStackNavigatorConfig } from "../../../navigation/navigatorConfig";
+import StepHeader from "../../../components/StepHeader";
+import { MethodSelection } from "./MethodSelection";
+import { EditTransactionParamList } from "../EditTransactionParamList";
+import SelectDevice from "../../../screens/SelectDevice";
+import SendConnectDevice from "../../../screens/ConnectDevice";
+import SendSummary from "../../../screens/SendFunds/04-Summary";
+import SendValidationSuccess from "../../../screens/SendFunds/07-ValidationSuccess";
+import SendValidationError from "../../../screens/SendFunds/07-ValidationError";
+import { TransactionAlreadyValidatedError } from "./TransactionAlreadyValidatedError";
 
 const Stack = createStackNavigator<EditTransactionParamList>();
